Playtest Changes
Updated cupboard interactable indicator to be more noticeable by changing the colour.
Increased the time limit in the level from 4-7 minutes as 4 was too short.
Added multiple instances of the win and lose screen GUIs so that even if a player was turned around, they dont just see a black void.